# full-stack-asset-transfer-guide/applications/rest-api/src/coconiko/coconiko_swagger.yaml
openapi: 3.0.1
info:
  title: Coconiko API
  version: 1.1.0
  description: |
    FabricネットワークにおけるユーザーとCoconikoのコインとNFTなどを管理するためのAPI。

    アプリからの呼び出しに最適化しており、ヘッダーに`X-API-Key`とログインしている`userId`を設定し、
    APIを呼び出す仕様となっています。
    - `X-Api-Key`で呼び出し元のシステムを識別しています
    - `userId`でログインしているユーザーを識別しています

    各APIの必要な認証情報は各APIの右側のロックアイコンで確認できます
    
    注意：「ユーザーID」と「アカウントID」が違います。
    - 「アカウントID」はFabricネットワーク上でユーザーを識別するIDです、新規ユーザーを登録する時に割り当てられます。`0x0`を保留しています。
    - 「ユーザーID」はアプリ側でユーザーを識別するIDです
...
paths:
...
  /api/coconiko/user:
    put:
      tags:
        - Coconiko
      summary: 新規ユーザーを登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - role
              properties:
                username:
                  type: string
                role:
                  type: string
                  description: ユーザーの役割
                  enum:
                    - user
                    - admin
              example:
                username: "test"
                role: "user"
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_name:
                    type: string
                    description: ユーザー名
                  app_id:
                    type: string
                    description: アプリケーションID
                  account:
                    type: object
                    description: アカウント情報
                    properties:
                      id:
                        type: string
                        description: アカウントID
                      role:
                        type: string
                        description: ユーザーの役割
                  timestamp:
                    type: string
                    format: date-time
                example:
                  user_name: "test"
                  app_id: "coconiko"
                  account:
                    id: "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                    role: "user"
                  timestamp: "2022-01-01T00:00:00.000Z"
        '201':
          description: ユーザーが存在しています
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
    get:
      tags:
        - Coconiko
      summary: ユーザー情報を取得
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    $ref: '#/components/schemas/UserInfo'
              example: {
                result: {
                  userId: "test",
                  accountid: "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                  role: "user",
                  balance: 0,
                  activate: true
                }
              }
    post:
      tags:
        - Coconiko
      summary: ユーザー情報を更新
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - active
              properties:
                active:
                  type: boolean
                  description: 有効
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    $ref: '#/components/schemas/UserInfo'
              example: {
                result: {
                  userId: "test",
                  accountid: "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                  role: "user",
                  balance: 0,
                  activate: true
                }
              }
  /api/coconiko/Mint:
    post:
      tags:
        - Coconiko Coin
      summary: コインを発行
      description: ログインしているユーザーによるコインの発行
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
                  description: コインの発行数量
                days:
                  type: integer
                  description: 有効期限（日数）、0にすると有効期限なし
                  default: 0
              example:
                amount: 100
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    description: "アカウントIDをキーとしてコインの残高を格納したオブジェクト"
                example:
                  result: {
                    "docType": "coconiko-coin",
                    "amount": 100,
                    "owner": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                    "burned": false
                  }
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/coconiko/BalanceOf:
    post:
      tags:
        - Coconiko Coin
      summary: 複数ユーザーの残高を取得
      description: 注意：アカウントIDが存在しない場合は、残高も0となります
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - owners
              properties:
                owners:
                  type: array
                  items:
                    type: string
                  description: アカウントIDの配列
                  example: [
                    "x509::/OU=client/CN=user1.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                    "x509::/OU=client/CN=user2.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                    "x509::/OU=client/CN=user3.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                    ]
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                example: {
                  result: {
                    "x509::/OU=client/CN=user1.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA": 100,
                    "x509::/OU=client/CN=user2.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA": 200,
                    "x509::/OU=client/CN=user3.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA": 300
                  }
                }
                properties:
                  result:
                    type: object
                    description: "アカウントIDをキーとしてコインの残高を格納したオブジェクト"
                    additionalProperties:
                      type: integer
                      description: コインの残高
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time          
  /api/coconiko/Transfer:
    post:
      tags:
        - Coconiko Coin
      summary: コインの転送
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - amount
              properties:
                to:
                  type: string
                  description: 転送先アカウントID
                  example: "x509::/OU=client/CN=user2.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                amount:
                  type: integer
                  description: コインの数量
                  example: 10
      security: 
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/coconiko/TransferFrom:
    post:
      tags:
        - Coconiko Coin
      summary: コインの転送（管理者用）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - to
                - amount
              properties:
                from:
                  type: string
                  description: 転送元アカウントID
                  example: "x509::/OU=client/CN=user1.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                to:
                  type: string
                  description: 転送先アカウントID
                  example: "x509::/OU=client/CN=user2.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                amount:
                  type: integer
                  description: 転送数量
                  example: 10
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
  /api/coconiko/TotalSupply:
    get:
      tags:
        - Coconiko Coin
      summary: コイン発行総量の取得
      description: |
        コイン発行総量の取得

        例（注意）：
        - 集計の開始日時が2024-12-27、集計の終了日時が2024-12-28の場合、集計期間が>=2024-12-28T00:00:00.000Z且つ<=2024-12-28T00:00:00.000Zになる
        - 集計の開始日時がnull、集計の終了日時が2024-12-28T10:10:10の場合、集計期間が<=2024-12-28T10:10:10.000Zになる
        - 集計の開始日時と集計の終了日時が設定されない場合は、集計期間が現時点までのレコードを集計する。
 
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          description: 集計の開始日時(UTC時間)
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          description: 集計の終了日時(UTC時間)
        - in: query
          name: activeUserOnly
          schema:
            type: boolean
            description: 有効ユーザーのみを集計するかどうか
            default: true
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      totalSupply:
                        type: integer
                        description: 発行されたコイン総量
                example: {
                  result: {
                    "totalSupply": 1000
                  }
                }
  /api/coconiko/BurnExpired:
    post:
      tags:
        - Coconiko Coin
      summary: 有効期限切れたコインを回収
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - owner
                - expirationDate
              properties:
                owner:
                  type: string
                  description: 所有者アカウントID
                expirationDate:
                  type: string
                  description: 有効期限
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 正常
...
  /api/coconiko/ClientAccountEventHistory/Count:
    get:
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          description: 開始日時(UTC時間)
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          description: 終了日時(UTC時間)
      tags:
        - Coconiko Coin
      summary: ユーザーのトランザクションイベント履歴の総数（指定期間）を取得する
      description: ""
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                example: 
                  {
                    "result": { count: 100}
                  }
  /api/coconiko/Summary:
    get:
      tags:
        - Coconiko Coin
      security:
        - ApiKeyAuth: []
          userId: []
      summary: 有効ユーザーの集計情報を取得する
      description: |
        有効ユーザーの集計情報を取得する（集計の開始日時 <= 集計期間 <= 集計の終了日時）

        例（注意）：
        - 集計の開始日時が2024-12-27、集計の終了日時が2024-12-28の場合、集計期間が>=2024-12-28T00:00:00.000Z且つ<=2024-12-28T00:00:00.000Zになる
        - 集計の開始日時がnull、集計の終了日時が2024-12-28T10:10:10の場合、集計期間が<=2024-12-28T10:10:10.000Zになる
        - 集計の開始日時と集計の終了日時が設定されない場合は、集計期間が現時点までのレコードを集計する。
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          description: 集計の開始日時(UTC時間)
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          description: 集計の終了日時(UTC時間)
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      totalMinted:
                        type: integer
                        description: 有効ユーザーの獲得コイン合計
                      totalUsed:
                        type: integer
                        description: 有効ユーザーの消費コイン合計
...
components:
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
        reason:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: object
    Result:
      type: object
      properties:
        result:
          type: object
          description: ページングされたクエリ結果
    UserInfo:
      type: object
      properties:
        userId:
          type: string
          description: ユーザーID
        accountId:
          type: string
          description: アカウントID
        role:
          type: string
          description: ユーザーの役割
        balance:
          type: integer
          description: ユーザーの残高
        active:
          type: boolean
          description: ユーザーの有効状態       
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: APPのAPIキー
    userId:
      type: apiKey
      in: header
      name: userId
      description: ログインしているユーザーID